+++ packages/analytics/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.0.40](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/analytics@7.0.39...@aws-amplify/analytics@7.0.40) (2024-07-23)
+
+**Note:** Version bump only for package @aws-amplify/analytics
+
+## [7.0.39](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/analytics@7.0.38...@aws-amplify/analytics@7.0.39) (2024-07-22)
+
+**Note:** Version bump only for package @aws-amplify/analytics
+
+## [7.0.38](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/analytics@7.0.37...@aws-amplify/analytics@7.0.38) (2024-07-19)
+
+**Note:** Version bump only for package @aws-amplify/analytics
+
+## [7.0.37](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/analytics@7.0.36...@aws-amplify/analytics@7.0.37) (2024-07-08)
+
+**Note:** Version bump only for package @aws-amplify/analytics
+
 ## [7.0.36](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/analytics@7.0.35...@aws-amplify/analytics@7.0.36) (2024-06-24)
 
 **Note:** Version bump only for package @aws-amplify/analytics