+++ packages/core/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.3.7](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/core@6.3.6...@aws-amplify/core@6.3.7) (2024-07-23)
+
+**Note:** Version bump only for package @aws-amplify/core
+
+## [6.3.6](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/core@6.3.5...@aws-amplify/core@6.3.6) (2024-07-22)
+
+### Bug Fixes
+
+- **aws-amplify:** createKeyValueStorageFromCookieStorageAdapter misses default path and secure values ([#13508](https://github.com/aws-amplify/amplify-js/issues/13508)) ([60a559f](https://github.com/aws-amplify/amplify-js/commit/60a559f612e092c76e1499f93547248312db46f2))
+
+## [6.3.5](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/core@6.3.4...@aws-amplify/core@6.3.5) (2024-07-19)
+
+**Note:** Version bump only for package @aws-amplify/core
+
+## [6.3.4](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/core@6.3.3...@aws-amplify/core@6.3.4) (2024-07-08)
+
+**Note:** Version bump only for package @aws-amplify/core
+
 ## [6.3.3](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/core@6.3.2...@aws-amplify/core@6.3.3) (2024-06-24)
 
 **Note:** Version bump only for package @aws-amplify/core