+++ packages/pubsub/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.1.15](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/pubsub@6.1.14...@aws-amplify/pubsub@6.1.15) (2024-07-23)
+
+**Note:** Version bump only for package @aws-amplify/pubsub
+
+## [6.1.14](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/pubsub@6.1.13...@aws-amplify/pubsub@6.1.14) (2024-07-22)
+
+**Note:** Version bump only for package @aws-amplify/pubsub
+
+## [6.1.13](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/pubsub@6.1.12...@aws-amplify/pubsub@6.1.13) (2024-07-19)
+
+**Note:** Version bump only for package @aws-amplify/pubsub
+
+## [6.1.12](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/pubsub@6.1.11...@aws-amplify/pubsub@6.1.12) (2024-07-08)
+
+**Note:** Version bump only for package @aws-amplify/pubsub
+
 ## [6.1.11](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/pubsub@6.1.10...@aws-amplify/pubsub@6.1.11) (2024-06-24)
 
 **Note:** Version bump only for package @aws-amplify/pubsub