+++ packages/react-native/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.4](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/react-native@1.1.3...@aws-amplify/react-native@1.1.4) (2024-07-23)
+
+**Note:** Version bump only for package @aws-amplify/react-native
+
+## [1.1.3](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/react-native@1.1.2...@aws-amplify/react-native@1.1.3) (2024-07-19)
+
+**Note:** Version bump only for package @aws-amplify/react-native
+
+## [1.1.2](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/react-native@1.1.1...@aws-amplify/react-native@1.1.2) (2024-07-08)
+
+**Note:** Version bump only for package @aws-amplify/react-native
+
 ## [1.1.1](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/react-native@1.1.0...@aws-amplify/react-native@1.1.1) (2024-05-13)
 
 **Note:** Version bump only for package @aws-amplify/react-native